+++ packages/nns/src/icp.ts
@@ -1,9 +1,12 @@
 import { ICPTs } from "../proto/ledger_pb";
 import type { FromStringToTokenError } from "./enums/token.enums";
-import { convertStringToE8s, ICPToken, TokenAmount } from "./token";
+import { convertStringToE8s, ICPToken, type Token } from "./token";
 
+// We don't extend to keep `fromE8s` and `fromString` as backwards compatible.
 // @deprecated
-export class ICP extends TokenAmount {
+export class ICP {
+  private constructor(protected e8s: bigint, public token: Token) {}
+
   public static fromE8s(amount: bigint): ICP {
     return new ICP(amount, ICPToken);
   }